platform/ai_dialogue_service/requirements.txt
charset-normalizer
chronological==0.1.1
et-xmlfile==1.1.0
idna==3.4
loguru==0.6.0
numpy==1.23.4
openai==1.2.2
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
tornado==6.0.4
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions
urllib3==1.26.12
apscheduler==3.8.0
nacos-sdk-python==0.1.12
PyYAML
cacheout==0.13.1
clueai